ca marche

+++ End_Puzzle.py
@@ -11,7 +11,7 @@ def puzzle_solution_generator(size):
     max_x = size
     max_y = size
     for n in range(1, size * size):
-        puzzle_solution[y * size + x] = n
+        puzzle_solution[y * size + x] = int(n)
         indexes_list.append((y, x))
         if x + ix == max_x:
             ix = 0
@@ -30,5 +30,5 @@ def puzzle_solution_generator(size):
         x += ix
         y += iy
     puzzle_solution[y * size + x] = 0
-    puzzle_solution = puzzle_solution.reshape(size, size)
+    puzzle_solution = puzzle_solution.reshape(size, size).astype(int)
     return puzzle_solution, indexes_list
